move to year_tab_page class
